Add caching services and refactor project structure
Implemented caching services in the common project to manage cached data. The structure of the project has been refactored, moving repository-related files and parts of the bot service to the DataAccess project. Additional changes include the removal of "JoinedDate" from "BlinkGuilds" and adjusting the bot service to use "DiscordSocketClient" instead of "DiscordShardedClient".