Simplify error handling and clean up imports
The code removed unnecessary switch-case blocks in interaction handlers that were merely placeholders. These have been replaced with simple agnostic error handling intending to implement a more sophisticated method and removed unused imports in 'Program.cs'. The configuraitonBuilder was dropped as it is no longer necessary with the updated code.