Add API services and update existing interfaces
The API project has been added to the solution, defining core services such as GuildsController and DiscordStartupService. Existent interfaces in the Caching and DIExtensions namespaces have been updated with summaries and param tags to provide additional documentation for developers. Relevant configuration files and docker setup details are also added for the API project.