Add 'complete' and 'remove' functionalities to todo list
The code modifications involve adding two new functionalities to the todo list in the 'TodoModule'. Users can now mark a todo item as complete or permanently remove it. An additional update has also been made in 'EmbedHelpers' to check if fields exist before attempting to add them in order to prevent potential errors.