Refactor GetId method in TodoModule
The GetId method's return type was changed from `Task<ulong?>` to `ulong`, removing the nullable return and async nature of the method. If unable to parse the supplied ID, it now throws an ArgumentOutOfRangeException instead of responding with an error and returning null. This update also affects the Complete and Remove methods.