Add authentication functionality and improve application structure
The code has been refactored to add authentication functionality mainly by introducing a new AuthController and configuring OAuth in the program file. Moreover, this update also includes an overall improvement of the application structure. Unnecessary code was removed, and some code was moved to better-organized files and folders. A configuration class was also added for managing application settings.