Refactor configuration and services registration
The configuration object is now fetched once using the `GetAppConfiguration()` extension method. This object is passed instead of the IConfiguration object when adding data access and caching services to improve consistency and reduce redundant calls. The DataAccess `DIExtensions` namespace and directory has been renamed to `Extensions` to match the naming in other parts of the project.