Add claims mapping for user authentication
Claims mapping for user identifiers, such as id, username, and global name was added to improve the authentication process. The mapping was accomplished by including System.Security.Claims in Program.cs and AuthController.cs and adding functional mapping methods in Program.cs. This change enhances the way user identity is validated and provides a more secure and reliable authentication.