Update Todo list UI and implement repository methods
Updated the Todo list page to improve the UI, adding checkboxes, description and delete option for each Todo item. Also, implemented the `GetAsync`, `AddAsync`, `UpdateAsync`, and `DeleteAsync` methods in the `TodoHttpRepository` which were previously just throwing `NotImplementedException`. The `GetAsync` method signature in the `ITodoHttpRepository` interface was updated to return a nullable UserTodo.