Add Todo creation feature to Todo page
This commit includes necessary changes to implement a new feature which allows users to create Todos on the Todo page. In addition to adding a form for Todo creation, the return type of GetAsync method has been modified from IReadOnlyCollection to IEnumerable. Minor modifications are made in other places such as changing the route from "/todo" to "/Todo" and throwing an exception if adding a todo item fails.