Add form validation to Todo form
To improve user experience, form validation was added to the Todo form fields. The `OnSubmit` attribute was changed to `OnValidSubmit` and `DataAnnotationsValidator` was added. Also, field inputs were updated with corresponding `ValidationMessage` components. Error messages were also added in the UserTodoDto model for Label and Description when exceeding their max length or when required fields are left blank.