Create TodoList component and refactor code
A new TodoList component has been created to centralize the representation of the ToDo list. This significant refactor aims to simplify code readability and reuse by migrating the corresponding code from the Todo page to the new TodoList component. The Commit also includes minor changes to CreateTodoForm to accommodate the refactor.