Add MoveMessageModule and update GetFriendlyName method
This commit introduces a new MoveMessageModule which enables messages to be moved to different channels. The BlinkModuleBase was also revised to allow interaction response or follow-up, depending on the state of the original interaction. In addition, the GetFriendlyName method was refreshed to return the friendly name of any discord user rather than only the guild user. Two other minor changes include: adding HttpClient service in Program.cs and translating strings to unsigned long in all new StringExtensions.