Add Wordle game functionality to Blink3.Core and Blink3.DataAccess
This commit introduces a new feature, Wordle game, to Blink3.Core and the Blink3.DataAccess. It adds entities for a Wordle game, the words guessed in the game, and the letters in the Wordle words. It also updates BlinkDBContext to handle these new entities and adjusts the .sln.DotSettings file to include Wordle.