Add image generation for Wordle guesses
Expanded the Wordle game service to generate an image representing the player's guess. This was implemented using SixLabors.ImageSharp and SixLabors.Fonts libraries. FontPath config property was added to manage the font file used for the image text. In addition, the bot's user interface was updated to respond with the generated image.