Refactor WordKey model to use record type
The WordKey model has been refactored to use the record type instead of a struct. This simplifies the code by eliminating the need for manually overriding the Equals, and GetHashCode methods. Now, the Language and Text properties in the WordKey model are immutable, which is an improvement towards more efficient and safer code.