Add dedicated image generation for WordleGuess
This commit introduces distinct classes for image generation of WordleGuess. These include WordleGuessImageGenerator for image creation and WordleGuessImageGeneratorOptions for defining parameters of the generated image. WordleLetterStateEnumExtensions class is also introduced for getting the relevant icon for each letter state. This change helps improve the code structure, making it clearer and more organized.