Update Docker workflow and expand README content
The Docker workflow has been updated to trigger on release creation instead of a pull request. A new tag mechanism is also added to distinguish between the branch and tag events. In addition, there have been extensive amendments to the README file to present the project structure, features, and build instructions for the newly renamed 'Blink3' project.