Integrate DiscordSocketClient into controllers
The controllers TodoController, GuildsController, and ApiControllerBase have been adjusted to include DiscordSocketClient as part of their constructor parameters. The Blink3 project file has also been updated to include additional package references for the integration of Discord functionality. More specifically, new logic was added in the 'InitDiscordClientAsync' method of the ApiControllerBase to wait for the DiscordBotClient to be ready before proceeding.